Factor out common dracoCompression methods and fields
