+++ packages/dev/core/src/Meshes/Compression/index.ts
@@ -1,2 +1,3 @@
 export * from "./dracoCompression";
 export * from "./meshoptCompression";
+export * from "./dracoDecoder";
